+++ src/views/HelpAndKnowledge/SolutionsHelp/_components/SolutionCategories/index.test.tsx
